Check if `mkl` has attribute `get_version_string`
It just happened on an installation of mine that import of a package which scooby failed, because of a 
```
AttributeError: module 'mkl' has no attribute 'get_version_string'
```
So it was able to import `mkl`, but `mkl` had no version info. This fix ensures that that won't make scooby fail.